fix: One component per file for TimePicker (#437)
* Rename file

* Rename temp

* Restore original

* Changes to refactor TimePicker into separate source files

* Remove script file

* Exported shared constant
